se agrego las ventanas de sweetalert

+++ js/admin.js
@@ -78,8 +78,12 @@ function agregarProducto() {
   limpiarFormulario();
   // cargar el producto nuevo en la fila de la tabla
   crearFilas(productoNuevo);
-
   //mostrar un mensaje al usuario
+  Swal.fire(
+    "Producto agregado",
+    "Su producto fue correctamente cargado",
+    "success"
+  );
 }
 
 function limpiarFormulario() {
@@ -151,31 +155,61 @@ function actualizarProducto() {
   listaProductos[posicionProducto].descripcion = descripcion.value;
   listaProductos[posicionProducto].url = url.value;
   // modificar el localstorage
-  localStorage.setItem('arregloProductos', JSON.stringify(listaProductos))
+  localStorage.setItem("arregloProductos", JSON.stringify(listaProductos));
   // volver a dibujar la tabla
   borrarFilas();
-  listaProductos.forEach((itemProducto)=>{ crearFilas(itemProducto) })
-// limpiar formulario
- limpiarFormulario();
+  listaProductos.forEach((itemProducto) => {
+    crearFilas(itemProducto);
+  });
+  // limpiar formulario
+  limpiarFormulario();
+  //  mostrar mensaje al usuario
+  Swal.fire(
+    "Producto modificado",
+    "Su producto fue correctamente modificado",
+    "success"
+  );
 }
 
-
-function borrarFilas(){
+function borrarFilas() {
   let tabla = document.querySelector("#tablaProducto");
-  tabla.innerHTML= '';
+  tabla.innerHTML = "";
 }
 
 window.eliminarProducto = (codigo) => {
-  console.log(codigo);
-  // aqui borramos el producto dentro del arreglo
-  let productosFiltrado = listaProductos.filter((itemProducto)=>{ return itemProducto.codigo != codigo });
+  Swal.fire({
+    title: "¿Esta seguro de eliminar este producto?",
+    text: "Una vez eliminado el producto no se puede recuperar el producto",
+    icon: "warning",
+    showCancelButton: true,
+    confirmButtonColor: "#3085d6",
+    cancelButtonColor: "#d33",
+    confirmButtonText: "Borrar producto",
+    cancelButtonText: "Cancelar",
+  }).then((result) => {
+    if (result.isConfirmed) {
+      // agregar todo el codigo que borra el producto
+      console.log(codigo);
+      // aqui borramos el producto dentro del arreglo
+      let productosFiltrado = listaProductos.filter((itemProducto) => {
+        return itemProducto.codigo != codigo;
+      });
 
-  console.log(productosFiltrado);
-  // actualizar el arreglo listaProductos
-  listaProductos = productosFiltrado;
-  // actualizo el localstorage
-  localStorage.setItem('arregloProductos', JSON.stringify(listaProductos))
-  // dibujar nuevamente la tabla
- borrarFilas();
- listaProductos.forEach((itemProducto)=>{ crearFilas(itemProducto)})
-}
+      console.log(productosFiltrado);
+      // actualizar el arreglo listaProductos
+      listaProductos = productosFiltrado;
+      // actualizo el localstorage
+      localStorage.setItem("arregloProductos", JSON.stringify(listaProductos));
+      // dibujar nuevamente la tabla
+      borrarFilas();
+      listaProductos.forEach((itemProducto) => {
+        crearFilas(itemProducto);
+      });
+      Swal.fire(
+        "Producto eliminado",
+        "El producto fue correctamente eliminado",
+        "success"
+      );
+    }
+  });
+};
